+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 # Changelog
+* Terraform rewrite
+
 * First release